shop/requirements.txt
Django==4.2
djangorestframework
psycopg2
Pillow
django-filters
django-cors-headers
requests
django-environ
